requirements.txt
opencv-python==4.5.5.64
Flask==2.0.3
Pillow==9.0.1
Werkzeug==2.0.3
colorthief==0.2.1